purchase page updated

+++ pages/purchase/index.php
@@ -1,6 +1,8 @@
 <?php
-require '../../includes/init.php';
-$purchases = select("SELECT Purchase.Id, Purchase.Quantity, BranchDetails.Id AS 'BranchDetailsId', Products.Id AS 'ProductId' FROM Purchase INNER JOIN BranchDetails ON Purchase.BranchId = BranchDetails.Id INNER JOIN Products ON Purchase.ProductId = Products.Id");
+require ('../../includes/init.php');
+$UserId = $_SESSION['UserId'];
+$permissions = authenticate('Purchase', $UserId);
+$purchases = select("SELECT Purchase.Id, Purchase.Quantity, BranchDetails.OwnerName AS 'BranchDetailsOwnerName', Products.Name AS 'ProductName' FROM Purchase INNER JOIN BranchDetails ON Purchase.BranchId = BranchDetails.Id INNER JOIN Products ON Purchase.ProductId = Products.Id");
 $index = 0;
 include pathOf('includes/header.php');
 include pathOf('includes/sidebar.php');
@@ -13,9 +15,11 @@ include pathOf('includes/sidebar.php');
           <div class="card-body">
             <div class="row justiyfy-content-between">
               <h4 class="card-title col-10">Purchase</h4>
-              <a class="btn btn-primary col-1 mb-5" href="./add.php">
-                <i class="mdi mdi-plus"></i>
-              </a>
+              <?php if ($permissions['AddPermission'] == 1) { ?>
+                <a class="btn btn-primary col-1 mb-5" href="./add">
+                  <i class="mdi mdi-plus"></i>
+                </a>
+              <?php } ?>
             </div>
             <div class="table-responsive">
               <table class="table">
@@ -25,33 +29,46 @@ include pathOf('includes/sidebar.php');
                     <th>Branch</th>
                     <th>Product</th>
                     <th>Quantity</th>
-                    <th>Modify</th>
-                    <th>Delete</th>
+                    <?php if ($permissions['EditPermission'] == 1) { ?>
+                      <th>Modify</th>
+                    <?php } ?>
+                    <?php if ($permissions['DeletePermission'] == 1) { ?>
+                      <th>Delete</th>
+                    <?php } ?>
                   </tr>
                 </thead>
                 <tbody>
-                  <?php foreach ($purchases as $purchase): ?>
-                    <tr>
-                      <td><?= $index += 1 ?></td>
-                      <td><?= $purchase['BranchDetailsId'] ?></td>
-                      <td><?= $purchase['ProductId'] ?></td>
-                      <td><?= $purchase['Quantity'] ?></td>
-                      <form action="./update.php" method="post">
-                        <td>
-                          <input type="hidden" name="Id" id="Id" value="<?= $purchase['Id'] ?>">
-                          <button type="submit" class="btn btn-primary btn-circle mb-2">
-                            <i class="mdi mdi-table-edit"></i>
-                          </button>
-                        </td>
-                      </form>
-                      <td>
-                        <button type="button" class="btn btn-primary btn-circle mb-2"
-                          onclick="deletePurchase(<?= $purchase['Id'] ?>)">
-                          <i class="mdi mdi-delete-variant"></i>
-                        </button>
-                      </td>
-                    </tr>
-                  <?php endforeach; ?>
+                  <?php if ($permissions['ViewPermission'] == 1) {
+
+                    foreach ($purchases as $purchase): ?>
+                      <tr>
+                        <td><?= $index += 1 ?></td>
+                        <td><?= $purchase['BranchDetailsOwnerName'] ?></td>
+                        <td><?= $purchase['ProductName'] ?></td>
+                        <td><?= $purchase['Quantity'] ?></td>
+                        <?php if ($permissions['EditPermission'] == 1) { ?>
+
+                          <form action="./update" method="post">
+                            <td>
+                              <input type="hidden" name="Id" id="Id" value="<?= $purchase['Id'] ?>">
+                              <button type="submit" class="btn btn-primary btn-circle mb-2">
+                                <i class="mdi mdi-table-edit"></i>
+                              </button>
+                            </td>
+                          </form>
+                        <?php } ?>
+                        <?php if ($permissions['DeletePermission'] == 1) { ?>
+
+                          <td>
+                            <button type="button" class="btn btn-primary btn-circle mb-2"
+                              onclick="deletePurchase(<?= $purchase['Id'] ?>)">
+                              <i class="mdi mdi-delete-variant"></i>
+                            </button>
+                          </td>
+                        <?php } ?>
+                      </tr>
+                    <?php endforeach;
+                  } ?>
                 </tbody>
               </table>
             </div>
@@ -60,26 +77,30 @@ include pathOf('includes/sidebar.php');
       </div>
     </div>
   </div>
- <?php
-    include pathOf('includes/footer.php');
-    include pathOf('includes/script.php');
-    ?>
-    <script>
-        function deletePurchase(Id) {
-            if (confirm("sure you want to delete this purchase"));
-            $.ajax({
-                url: "../../api/purchase/delete.php",
-                method: "POST",
-                data: {
-                    Id: Id
-                },
-                success: function (response) {
-                    alert('Purchase Deleted!');
-                    window.location.href = './index.php';
-                }
-            })
-        }
-    </script>
-<?php
-include pathOf('includes/pageEnd.php');
-?>
+  <?php
+  include pathOf('includes/footer.php');
+  include pathOf('includes/script.php');
+  ?>
+
+  <script>
+
+    function deletePurchase(Id) {
+      if (confirm("Are you sure you want to delete this Purchase?")) {
+        $.ajax({
+          url: "../../api/purchase/delete",
+          method: "POST",
+          data: {
+            Id: Id
+          },
+          success: function (response) {
+            alert('Purchase  deleted!');
+            window.location.href = './index';
+          }
+        });
+      }
+    }
+
+  </script>
+  <?php
+  include pathOf('includes/pageEnd.php');
+  ?>
